requirements.txt
Flask
Flask-SocketIO
python-dotenv
supabase
PyJWT
requests
eventlet
gunicorn
qrcode[pil]

